Fix race condition error
There is an odd race condition that causes the dest path to not be
available when either browserify or metalsmith tasks are ran at the same
time. We need to provide a dependency to delay this task till after the
previous two tasks are complete. However calling the styles task
directly does not need those dependencies and so we provide two tasks.
An internal one and a user facing one since they both implement the same
task.

The error was an IO error (unable to stat) presumably due to possible
file locking?