Add a inPlace template pre filter plugin
The metalsmith-templates plugin has an [inPlace][1] option useful for
interpolating / rendering documents as if they themselves were
templates. However, this option assumes all files passed in are
appropriately typed for the template engine in question. This means
files without a template property or non template-able files (images)
will break most template engines.

To combat this issue this plugin will work just like the normal
metalsmith-template but pre-filter the files it uses based on if it has
a template property.

[1]: https://github.com/segmentio/metalsmith-templates/commit/dcf84028