Add partial support to pretemplates plugin
Seems when used with inPlace it will ignore partial loading like normal
forcing the partials option to be prepopuated with the file contents.
prepopuated now taks the whole filename and reads in the contents. When
callin pretemplates pass in the template options with the partials
object values to resolve-able file paths.

This is in contrast to the normal metalsmith-templates plugin which
expects the partials option to be an object with name and file basename
no extention.